config: separate hash library and improve some details
