+++ dvc/config.py
@@ -3,6 +3,7 @@
 from __future__ import unicode_literals
 
 from dvc.utils.compat import str, open, urlparse
+from dvc.utils import hash
 
 import os
 import errno
@@ -58,12 +59,9 @@ def supported_hash_local(hash_names):
     Args:
         hash_names (list/string): hash name(s).
     """
-    if isinstance(hash_names, str):
-        hash_names = [h.strip() for h in hash_names.split(",")]
-    for h in hash_names:
-        if h not in ["sha256", "md5"]:
-            return False
-    return True
+    if hash.checksum_types_from_str(hash_names):
+        return True
+    return False
 
 
 def supported_cloud(cloud):
