+++ dvc/output/__init__.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 import schema
 
-from dvc import utils
+from dvc.utils import hash
 from dvc.utils.compat import urlparse, str
 
 from dvc.output.base import OutputBase
@@ -41,8 +41,8 @@ OUTS_MAP = {
 # so when a few types of outputs share the same name, we only need
 # specify it once.
 CHECKSUM_SCHEMA = {
-    schema.Optional(utils.CHECKSUM_MD5): schema.Or(str, None),
-    schema.Optional(utils.CHECKSUM_SHA256): schema.Or(str, None),
+    schema.Optional(hash.CHECKSUM_MD5): schema.Or(str, None),
+    schema.Optional(hash.CHECKSUM_SHA256): schema.Or(str, None),
     schema.Optional(RemoteS3.PARAM_CHECKSUM): schema.Or(str, None),
     schema.Optional(RemoteHDFS.PARAM_CHECKSUM): schema.Or(str, None),
 }
