+++ dvc/remote/local/__init__.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 from copy import copy
 
 from dvc.remote.local.slow_link_detection import slow_link_guard
-from dvc import utils
+from dvc.utils import hash
 from dvc.utils.compat import str, makedirs
 
 import os
@@ -22,15 +22,8 @@ from dvc.remote.base import (
     STATUS_DELETED,
     STATUS_MISSING,
 )
-from dvc.utils import (
-    remove,
-    move,
-    copyfile,
-    to_chunks,
-    tmp_fname,
-    file_checksum,
-    walk_files,
-)
+from dvc.utils import remove, move, copyfile, to_chunks, tmp_fname, walk_files
+from dvc.utils.hash import file_checksum
 from dvc.config import Config
 from dvc.exceptions import DvcException
 from dvc.progress import progress
@@ -75,14 +68,14 @@ class RemoteLOCAL(RemoteBase):
         else:
             self.cache_types = copy(self.DEFAULT_CACHE_TYPES)
 
-        self.hash = [utils.CHECKSUM_MD5]
+        self.hash = [hash.CHECKSUM_MD5]
         if repo and repo.config and repo.config.config:
             conf = repo.config.config[Config.SECTION_HASH]
             hash_local = conf.get(Config.SECTION_HASH_LOCAL, None)
             if hash_local:
-                if isinstance(hash_local, str):
-                    hash_local = [h.strip() for h in hash_local.split(",")]
-                self.hash = hash_local
+                self.hash = (
+                    hash.checksum_types_from_str(hash_local) or self.hash
+                )
 
         if self.cache_dir is not None and not os.path.exists(self.cache_dir):
             os.mkdir(self.cache_dir)
